Milestone 1 Complete
The project is ready for Milestone 1, now that the movement is working, I have a very basic scene with some simple objects hanging around, and added a crouch mechanic as the extra feature. I've also created some basic colored materials to have on the objects so we aren't looking at a basic and boring white scene.

Known bug: whenever the player object collides with another object the player starts to gradually increase on their Y-axis. Current cause is unknown and will look into it moving forward.