Catching Up
Updated the IKController, although there seems to be some issue with how the arms move, and they don't snap to the points correctly.  I also added machine gun, sniper rifle, uzi, and bullet models to the project to use for ammo and different unique weapons.  Working on making the machine gun a burst rifle, but due to some other errors with the IKController and the pickups, and possibly more, the game is currently not functional.