IK and Burst Rifle Fixes
Fixed an error in the weapon script that wasn't hooking things up correctly, and modified the equip and unequip system to have a null reference check first, and now the IK does exactly what it is supposed to.  I added a generic spawner to try and spawn weapon pickups, but they won't actually spawn so I need to go back and look through it again.