Added Health UI and Raycast Sniper
Made the sniper rifle now shoot on raycast and spawn a little star particle effect on both the firepoint and the object that is hit.
Also added a health bar system for the player and a target dummy to show each object having different health values.
Next step is to add a fire particle system as my extra feature for this Milestone.