Options Menu and Lighting Finished
I have created the options menu UI that runs off of the main menu scene which has sliders for the master, music, and SFX volumes, and a resolution and graphics dropdown. I also changed the directional light to a dark blue and added a point light on top of my fire particle system.